+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **[2.0.0]**
 - Added support for masked enemies! Masked enemies that spawn from dog players will now use the dog model and no masks
+- Added a new config file setting to disable masks on dog mimics
 
 **[1.1.4]**
 - Removed patch for adding blood decals to a dog player